Agrego manejo de errores Lexicos y sinctactios
1.Manejo de errores lexicos detallados
2.Manejo de errores sintactos preliminares
3.Se agrego scripts a index.js
4.Estructura archivos cambiada y agrupada por Ts y Js
