+++ Jison/XPATH/.gitignore
@@ -1,0 +1,2 @@
+/errores.txt
+/pruebas.txt
