--- XPATH/pruebas.txt
@@ -1,92 +1,0 @@
-/libreria/area1/libros
-|libro[child::existencias]
-
-|//libro[child::existencias]
-
-
-|/libreria/area1/libros/libro
-
-|//@sexo
-
-|//autor[@sexo="macho"]
-
-|/bookstore/book[title="Holi"]
-
-|/tengo/sueNo[@ganasDeSeguir="noHay"]
-
-|//title/parent::book
-
-|/juan/pedro/hijos
-
-|juan/casera/hijos_casera
-
-|autor[@sexo="macho"]
-
-|bookstore[@eshta = "grande"] | bookstore
-
-|bookstore/book/title/parent::book[title="asdsadasdasd"]
-
-|bookstore/book/title[ancestor::book]
-
-|bookstore/book/title[ancestor::book and child::asdasd]
-
-|bookstore/book/title[sda=12.50 and child::asdasd]
-
-|bookstore/book/title[2+5]
-
-|/./././bookstore
-
-|./bookstore/book/title
-
-|./bookstore[child::book]
-
-|/bookstore/book/title/..[@category="web"]
-
-|/bookstore/book/title[ancestor::book[title="Learning XML"]]
-
-|bookstore/book/title/parent::book[@category="web"]
-
-|./child::eshta[marta="juan"][pedro="juan"]
-
-|/bookstore/book[@category="web"][child::title[.="Learning XML"]]
-
-
-|//book/descendant::author[.="James"]
-
-|//book[descendant::author]
-
-|//book/author[.]
-
-|//book/author[author=("J")]
-
-|bookstore/book/title/.
-
-|bookstore/book/*
-
-|//@*
-
-|//@year[.=1969]/../../autor
-
-|../fechaPublicacion/@year
-
-|//titulo[../fechaPublicacion/@year=//@year[../../titulo="La vida está en otra parte"]]
-
-|//libro[autor=//libro[titulo="Pantaleón y las visitadoras"]/autor]/titulo
-
-|//autor[../fechaPublicacion/@year=1969]
-
-|//libro[autor=//libro[titulo="Pantaleón y las visitadoras"]/autor/textl]/titulo
-
-|/bookstore//node()
-
-|/bookstore/book/title/@node()
-
-|/bookstore/book[position()<3]
-
-|/bookstore/book[last()-1]
-
-|bookstore/book/child::*/child::price
-
-|bookstore/book/child::node()/child::price
-
-| /bookstore/book/title/self::title[.][.][.][.][.][.]/self::title[.][.][.][.][.][.][.]/self::title[.][.][.][.][.][.][.]/self::title[.][.][.][.][.][.][.]/self::title[.][.][.][.][.][.][.]/self::title[.][.][.][.][.][.][.]/self::title[.][.][.][.][.][.][.]/self::title[.][.][.][.][.][.][.]/self::title[.][.][.][.][.][.][.]/self::title[.][.][.][.][.][.][.]/self::title[.][.][.][.][.][.][.]/self::title[.][.][.][.][.][.][.]/self::title[.]/../.
